relocate store types
