+++ src/Types/index.ts
@@ -3,51 +3,6 @@
  * SPDX-License-Identifier: AGPL-3.0-or-later
  */
 
-export { Activity, Activities } from '../stores/activity.ts'
-export { UpdateType, Group, AppSettings } from '../stores/appSettings.ts'
-export { Combo } from '../stores/combo.ts'
-export { Comment, Comments, CommentsGrouped } from '../stores/comments.ts'
-export {
-	Poll,
-	PollType,
-	AccessType,
-	ShowResults,
-	AllowProposals,
-	PollConfiguration,
-	PollStatus,
-	PollPermissions,
-	CurrentUserStatus,
-} from '../stores/poll.ts'
-
-export {
-	SortType,
-	FilterType,
-	PollCategory,
-	Meta,
-	PollList,
-} from '../stores/polls.ts'
-
-export {
-	Option,
-	OptionVotes,
-	Sequence,
-	SimpleOption,
-	Options,
-} from '../stores/options.ts'
-
-export { Share, Shares, ShareType } from '../stores/shares.ts'
-export { Route, UserStatus, Session } from '../stores/session.ts'
-
-export {
-	UserPreferences,
-	SessionSettings,
-	Calendar,
-	Preferences,
-	ViewMode,
-} from '../stores/preferences.ts'
-
-export { Answer, AnswerSymbol, Vote, Votes } from '../stores/votes.ts'
-
 export {
 	DateTimeUnit,
 	DateTimeUnitType,
@@ -114,26 +69,6 @@ export type ISearchType = 0 | 1 | 2 | 4 | 7 | 51 | 99
 export type Chunking = {
 	size: number
 	loaded: number
-}
-
-export type ApiEmailAdressList = {
-	displayName: string
-	emailAddress: string
-	combined: string
-}
-
-export type AppPermissions = {
-	addShares: boolean
-	addSharesExternal: boolean
-	allAccess: boolean
-	changeForeignVotes: boolean
-	comboView: boolean
-	deanonymizePoll: boolean
-	pollCreation: boolean
-	pollDownload: boolean
-	publicShares: boolean
-	seeMailAddresses: boolean
-	unrestrictedOwner: boolean
 }
 
 export type User = {
