+++ src/stores/appSettings.ts
@@ -4,61 +4,15 @@
  */
 
 import { defineStore } from 'pinia'
+
 import { AppSettingsAPI } from '../Api/index.ts'
 import { Logger } from '../helpers/index.ts'
-import { AxiosError } from '@nextcloud/axios'
 
-export type UpdateType = 'noPolling' | 'periodicPolling' | 'longPolling'
-export type Group = {
-	id: string
-	userId: string
-	displayName: string
-	emailAddress: string
-	isNoUser: boolean
-	type: string
-}
-
-export type AppSettings = {
-	allAccessGroups: string[]
-	allowCombo: boolean
-	allowPublicShares: boolean
-	allowAllAccess: boolean
-	allowPollCreation: boolean
-	allowPollDownload: boolean
-	autoArchive: boolean
-	autoArchiveOffset: number
-	autoDelete: boolean
-	autoDeleteOffset: number
-	defaultPrivacyUrl: string
-	defaultImprintUrl: string
-	disclaimer: string
-	imprintUrl: string
-	legalTermsInEmail: boolean
-	privacyUrl: string
-	showMailAddresses: boolean
-	showLogin: boolean
-	unrestrictedOwner: boolean
-	updateType: UpdateType
-	useActivity: boolean
-	useCollaboration: boolean
-	useSiteLegalTerms: boolean
-	navigationPollsInList: boolean
-	finalPrivacyUrl: string
-	finalImprintUrl: string
-	comboGroups: string[]
-	publicSharesGroups: string[]
-	pollCreationGroups: string[]
-	pollDownloadGroups: string[]
-	showMailAddressesGroups: string[]
-	unrestrictedOwnerGroups: string[]
-	groups: Group[]
-	status: {
-		loadingGroups: boolean
-	}
-}
+import type { AxiosError } from '@nextcloud/axios'
+import type { AppSettingsStore } from './appSettings.types.ts'
 
 export const useAppSettingsStore = defineStore('appSettings', {
-	state: (): AppSettings => ({
+	state: (): AppSettingsStore => ({
 		allAccessGroups: [],
 		allowCombo: true,
 		allowPublicShares: true,
