+++ src/stores/options.ts
@@ -4,61 +4,26 @@
  */
 
 import { defineStore } from 'pinia'
-import { PublicAPI, OptionsAPI } from '../Api/index.ts'
-import { User } from '../Types/index.ts'
-import { Logger } from '../helpers/index.ts'
 import orderBy from 'lodash/orderBy'
+
+import { PublicAPI, OptionsAPI } from '../Api/index.ts'
+import { Logger } from '../helpers/index.ts'
+
 import { usePollStore } from './poll.ts'
 import { useSessionStore } from './session.ts'
-import { Answer, useVotesStore } from './votes.ts'
-import { DateTimeUnitType, TimeUnitsType } from '../constants/dateUnits.ts'
-import { AxiosError } from '@nextcloud/axios'
+import { useVotesStore } from './votes'
 
-export type RankedType = 'yes' | 'no'
-
-export type Sequence = {
-	unit: DateTimeUnitType
-	stepWidth: number
-	repetitions: number
-}
-
-export type OptionVotes = {
-	yes: number
-	maybe: number
-	no: number
-	count: number
-	currentUser?: Answer
-}
-
-export type SimpleOption = {
-	text?: string
-	timestamp?: number
-	duration?: number
-}
-
-export type Option = {
-	id: number
-	pollId: number
-	text: string
-	timestamp: number
-	deleted: number
-	order: number
-	confirmed: number
-	duration: number
-	locked: boolean
-	hash: string
-	isOwner: boolean
-	votes: OptionVotes
-	owner: User | undefined
-}
-
-export type Options = {
-	options: Option[]
-	ranked: RankedType
-}
+import type { AxiosError } from '@nextcloud/axios'
+import type { TimeUnitsType } from '../constants/dateUnits.ts'
+import type {
+	Sequence,
+	SimpleOption,
+	Option,
+	OptionsStore,
+} from './options.types.ts'
 
 export const useOptionsStore = defineStore('options', {
-	state: (): Options => ({
+	state: (): OptionsStore => ({
 		options: [],
 		ranked: 'no',
 	}),
