+++ src/stores/pollGroups.types.ts
@@ -2,7 +2,7 @@
  * SPDX-FileCopyrightText: 2025 Nextcloud contributors
  * SPDX-License-Identifier: AGPL-3.0-or-later
  */
-import { User } from '../Types/index.ts'
+import type { User } from '../Types/index.ts'
 
 export type PollGroup = {
 	id: number
